requirements.txt
discord.py==2.5.2
openai==1.75.0
python-dotenv==1.1.0
Flask==3.1.0
datasette==0.64.3
sqlite-utils==3.36
wandb==0.19.9
